+++ src/components/apple-auth.tsx
@@ -13,7 +13,7 @@ import { Skeleton } from "./ui/skeleton"
 
 const AppleAuth = ({ loading }: { loading: boolean }) => {
   const clientId = env.NEXT_PUBLIC_APPLE_OAUTH_CLIENT_ID
-  const redirectURI = env.NEXT_PUBLIC_APP_URL
+  const redirectURI = `${env.NEXT_PUBLIC_APP_URL}/oauth-callback`
 
   const { authIframeClient } = useTurnkey()
   const { loginWithOAuth } = useAuth()
@@ -21,16 +21,6 @@ const AppleAuth = ({ loading }: { loading: boolean }) => {
   const [nonce, setNonce] = useState<string>("")
   const [storedToken, setStoredToken] = useState<string | null>(null) // Store the token locally
   const [hasLoggedIn, setHasLoggedIn] = useState(false) // Track if loginWithOAuth has been called
-
-  const searchParams = useSearchParams()
-
-  // Get token from query string params and store in state when available
-  useEffect(() => {
-    const token = searchParams.get("id_token")
-    if (token) {
-      setStoredToken(token) // Store token if available
-    }
-  }, [searchParams])
 
   // Generate nonce based on iframePublicKey
   useEffect(() => {
@@ -43,31 +33,13 @@ const AppleAuth = ({ loading }: { loading: boolean }) => {
     }
   }, [authIframeClient?.iframePublicKey])
 
-  // Trigger loginWithOAuth when both token and iframePublicKey are available, but only once
-  useEffect(() => {
-    if (storedToken && authIframeClient?.iframePublicKey && !hasLoggedIn) {
-      console.log("doing the thing to the things")
-
-      // Call the OAuth login function with the stored token
-      loginWithOAuth(storedToken)
-
-      // Set flag to prevent further calls
-      setHasLoggedIn(true)
-    }
-  }, [
-    storedToken,
-    authIframeClient?.iframePublicKey,
-    hasLoggedIn,
-    loginWithOAuth,
-  ])
-
   return (
     <>
       {nonce ? (
         <AppleLogin
           clientId={clientId}
           redirectURI={redirectURI as string}
-          responseType="id_token code"
+          responseType="code id_token"
           nonce={nonce}
           responseMode="fragment"
         />
